timing met
There was initially a problem with closure on setup timing (negative setup slack) for CLK_25.
That was due to a long arithmatic path that had to happen in one clock cycle (multiplication, divison and addition).
There are two solutions in this commit: procedure for sample scaling that approximates the division by a fixed-point reciprocal multiplication or a new module (sampleToScreenPipeline.vhd) that pipelines the three arithmatic steps.
Both solutions are able to close timing.
In the pipelined solution, the division by 50 needed to be replaced by the approximated fixed-point reciprocal multiplication as well, because the pure division by 50 would take 3 clock cycles.